new attempt at rust persistent worker: just the build setting for now
going to try this in C++ and have it all work nicely.
This commit just sets up a build setting, so that users can customize
whether to use the worker by just setting the flag on the command line
or in a local bazelrc file. This allows easily switching between
worker and non-worker.

Use as `bazel build <rust target> --@rules_rust//rust:use-worker`.
Obviously this will fail right now.